feat:finished stream logic in home and connect to order flow

+++ presentation/src/main/kotlin/com/medsy/presentation/offers/components/PriceRow.kt
@@ -14,7 +14,7 @@ import com.medsy.presentation.R
 import com.medsy.presentation.common.util.PriceFormatter
 
 @Composable
-fun PriceRow(label: String, price: Int) {
+fun PriceRow(label: String, price: Double) {
     val locale = LocalConfiguration.current.locales[0]
     Row(
         modifier = Modifier.fillMaxWidth(),
